Multiplication Table Generator

+++ control-flow/multiplication_table.py
@@ -1,6 +1,6 @@
 #multiplication_table.py
 #Prompt User for a Number:
-number = float(input("Enter a number to see its multiplication table:")
+number =int(input("Enter a number to see its multiplication table:")
 #Generate and Print the Multiplication Table 1 to 10:
 for i range(1, 11):
     result = number * i
